+++ digits/changelog.md
@@ -2,9 +2,10 @@
 *Non-trivial pull requests should include an entry below. Entries must be suitable for inclusion in public-facing materials such as release notes and blog posts. Keep them short, sweet, and in the past tense. New entries go on top. When merging to deploy, add the version number and date.*
 
 ## Unreleased
-* (IC) Add weak reference to hold the AuthCallback from the developer in LoginResultReceiver
+* (EF) Fixed critical issue where Digits sessions are lost when using Proguard.
 
 ## v1.6.1
+* (IC) Add weak reference to hold the AuthCallback from the developer in LoginResultReceiver
 
 ## v1.6.0
 * (IC) [Fixed crash on StateButton when multiple clicks] (https://github